+++ app/javascript/form/face_preview.js
@@ -1,0 +1,3 @@
+document.addEventListener('DOMContentLoaded', function(){
+  console.log(123);
+});
