+++ app/javascript/packs/application.js
@@ -7,12 +7,13 @@ require("@rails/ujs").start()
 // require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
-require("modal_window/employee_create")
-require("modal_window/worker_create")
-require("modal_window/chat_create")
-require("modal_window/customer_create")
-require("modal_window/customer_meeting_create")
-require("modal_window/customer_construction_create")
+require("form/employee_create")
+require("form/worker_create")
+require("form/face_preview")
+require("form/chat_create")
+require("form/customer_create")
+require("form/customer_meeting_create")
+require("form/customer_construction_create")
 require("customer/status_color")
 
 // Uncomment to copy all static images under ../images to the output folder and reference
