+++ app/views/shared/form/user/_employee.html.erb
@@ -19,6 +19,11 @@
 
         <%# 顔写真 %>
         <div class="user-sign--card--info--child">
+          <div id="image-preview"></div>
+        </div>
+
+        <%# 顔写真 %>
+        <div class="user-sign--card--info--child">
           <%= f.label :face_image, '顔写真', class:"user-sign--card--info--child--tag" %>
           <%= f.file_field :face_image, class:"user-sign--card--info--child--value image-select" %>
         </div>
